Refactored hierarchy to run only when group is updated
